+++ Client/src/components/Profile/Tribute.jsx
@@ -7,11 +7,13 @@ export default function Tribute({ tribute, onDelete }) {
   const user = JSON.parse(localStorage.getItem("user"));
 
   const handleDeleteTribute = async () => {
-    try {
-      await deleteTribute(tribute.id);
-      onDelete(); // Call the onDelete prop to refresh the tributes list
-    } catch (error) {
-      console.error("Error deleting tribute:", error);
+    if(window.confirm("Are you sure you want to delete this post? This action cannot be undone.")){
+      try {
+        await deleteTribute(tribute.id);
+        onDelete(); // Call the onDelete prop to refresh the tributes list
+      } catch (error) {
+        console.error("Error deleting tribute:", error);
+      }
     }
   };
